Resolved  #6919. It might be required to rework unit tests.
